Fix OMake build on Mavericks
Fix dependency on readline

This commit fixes omake's building on my system. I do this with a lot of
assumptions, which I explain below.

OMake fails building because it tries to use GNU readline instead of BSD
readline.
* Autodetection tests for `rl_on_new_line` to determine which readline it
  is linking against. But `otool -t -v /usr/lib/libreadline.dylib` shows
  that this function is now defined by the system BSD readline. So it
  mistakenly detects that a GNU readline is available.
* Then, it tries to use function history_list from GNU libhistory and
  fails.

To fix this, we need to:

* also link `libhistory` in. I assume that libhistory was part of
  libreadline at some point, so I modified the detection routines to
  also link libhistory when detecting GNU's readline library.
* Have readline available at all during linking. Adding a dependency
  appears sufficient for the job.
* fix detection of GNU readline. Simply testing for history_list should
  do the job; I know that because linking OMake failed on *that*
  function. This will ensure that software using OMake correctly detects
  which readline library is available.

Note: On my system I do have GNU libreadline installed, but I don't
think that's what broke the build. The last confirmed build of this
library is from June 2013, before the release of OS X Mavericks (which I
use), so I suppose that rl_on_new_line was not part of OS X's
`libreadline` yet.

I've added a link to the upstream report.

Also, specify version in build. Autodetection fails here, leaving us
with version '1'.

Closes #22.

Signed-off-by: Adam Vandenberg <flangy@gmail.com>